refactor(execution-engine): extract task-runner-dispatch from task-runner
Moves executor selection and startup (dispatchExecutor: the SSH-failover retry loop, pre-start heartbeats, start-metadata persistence, active-execution registration and stream wiring) out of task-runner into task-runner-dispatch, behavior unchanged. Drops the launch-support and RunnerKind imports whose last inline users moved out.
One atomic slice of the step-4 large-file decomposition.
