+++ packages/execution-engine/src/task-runner.ts
@@ -12,7 +12,7 @@ import { randomUUID } from 'node:crypto';
 import { homedir } from 'node:os';
 
 import { scopePlanTaskId } from '@invoker/workflow-core';
-import type { Orchestrator, TaskState, ExperimentVariant, RunnerKind } from '@invoker/workflow-core';
+import type { Orchestrator, TaskState, ExperimentVariant } from '@invoker/workflow-core';
 import type { SQLiteAdapter } from '@invoker/data-store';
 import type { WorkRequest, WorkResponse, ActionType, Logger } from '@invoker/contracts';
 import type { Executor, ExecutorHandle } from './executor.js';
@@ -62,8 +62,9 @@ import {
 import { assertNotGitConfigMutation, ensureRemoteUrl } from './git-config-mutation.js';
 import { killProcessGroup, SIGKILL_TIMEOUT_MS } from './process-utils.js';
 import { retryTransientGitHubCli } from './git-utils.js';
-import { PRE_START_HEARTBEAT_INTERVAL_MS, getExecutorStartTimeoutMs, isRetryableSshStartupTransportError, nextLeaseExpiry, type StartupFailureMetadata } from './task-runner-launch-support.js';
+import { PRE_START_HEARTBEAT_INTERVAL_MS, nextLeaseExpiry } from './task-runner-launch-support.js';
 import { buildWorkRequest } from './task-runner-prepare.js';
+import { dispatchExecutor } from './task-runner-dispatch.js';
 
 export type { TaskHeartbeatEvent, TaskRunnerCallbacks } from './task-runner-callbacks.js';
 type ReviewGateState = NonNullable<TaskState['execution']['reviewGate']>;
@@ -722,330 +723,17 @@ export class TaskRunner {
 
     const request = await buildWorkRequest(this, { task, attemptId, bench });
 
-    const startGeneration = task.execution.generation ?? 0;
-    const actionType = this.determineActionType(task);
-    const executionAgent = task.config.executionAgent?.trim() || DEFAULT_EXECUTION_AGENT;
-
-    traceExecution(
-      `${RESTART_TO_BRANCH_TRACE} executeTaskInner taskId=${task.id} WorkRequest built actionType=${request.actionType} repoUrl=${request.inputs.repoUrl ?? '(none)'} upstreamBranches=${JSON.stringify(request.inputs.upstreamBranches ?? [])}`,
-    );
-    const startT0 = Date.now();
-    const attemptedPoolMemberKeys = new Set<string>();
-    let executor!: Executor;
-    let handle!: ExecutorHandle;
-    while (true) {
-      bench('selectExecutor.start');
-      executor = this.selectExecutor(task, attemptedPoolMemberKeys);
-      const poolSelectionForStart = this.pendingPoolSelections.get(task.id);
-      if (!this.acquirePoolSelectionLease(task, attemptId, poolSelectionForStart)) {
-        if (poolSelectionForStart) {
-          attemptedPoolMemberKeys.add(poolSelectionForStart.memberKey);
-          this.pendingPoolSelections.delete(task.id);
-        }
-        continue;
-      }
-      bench('selectExecutor.end', {
-        executorType: executor.type,
-      });
-      traceExecution(
-        `${RESTART_TO_BRANCH_TRACE} executeTaskInner taskId=${task.id} selectExecutor → type=${executor.type} calling executor.start()`,
-      );
-      traceExecution(`[trace] TaskRunner: task=${task.id} calling executor.start() type=${executor.type}`);
-      this.logger.info(
-        `[TaskRunner] executor.start begin task=${task.id} attempt=${attemptId} executor=${executor.type} ` +
-          `generation=${task.execution.generation ?? 0}`,
-      );
-      this.persistence.logEvent?.(task.id, 'task.executor.start_begin', {
-        dispatchId: dispatchOpts?.dispatchId,
-        attemptId,
-        executorType: executor.type,
-        poolId: poolSelectionForStart?.poolId,
-        poolMemberId: poolSelectionForStart?.member.id,
-      });
-      bench('onLaunchStart.before', {
-        executorType: executor.type,
-      });
-      this.callbacks.onLaunchStart?.(task.id, executor);
-      bench('executor.start.before', {
-        executorType: executor.type,
-      });
-      const startTimeoutMs = getExecutorStartTimeoutMs();
-      const preStartHeartbeatTimer = setInterval(() => {
-        const now = new Date();
-        this.renewPoolSelectionLease(poolSelectionForStart);
-        this.persistence.updateAttempt?.(attemptId, {
-          lastHeartbeatAt: now,
-          leaseExpiresAt: nextLeaseExpiry(now),
-        } as any);
-        this.callbacks.onHeartbeat?.(task.id, { at: now, source: 'executor' });
-      }, PRE_START_HEARTBEAT_INTERVAL_MS);
-      let preStartTimeout: ReturnType<typeof setTimeout> | undefined;
-      try {
-        handle = await Promise.race<ExecutorHandle>([
-          executor.start(request),
-          new Promise<ExecutorHandle>((_resolve, reject) => {
-            preStartTimeout = setTimeout(() => {
-              reject(new Error(`Executor startup timed out after ${startTimeoutMs}ms (${executor.type})`));
-            }, startTimeoutMs);
-          }),
-        ]);
-        break;
-      } catch (err) {
-        const meta = err as StartupFailureMetadata;
-        if (
-          executor.type === 'ssh'
-          && poolSelectionForStart?.member.type === 'ssh'
-          && !meta.workspacePath
-          && !meta.branch
-          && isRetryableSshStartupTransportError(err)
-        ) {
-          attemptedPoolMemberKeys.add(poolSelectionForStart.memberKey);
-          const pool = this.getExecutionPools()[poolSelectionForStart.poolId];
-          const hasAnotherSshMember = pool?.members.some((member) =>
-            member.type === 'ssh' && !attemptedPoolMemberKeys.has(this.poolMemberKey(member)),
-          ) ?? false;
-          if (hasAnotherSshMember) {
-            const retryMessage =
-              `Executor startup failed (${executor.type}) on pool member ${poolSelectionForStart.member.id}; ` +
-              `retrying another SSH pool member: ${err instanceof Error ? err.message : String(err)}\n`;
-            this.callbacks.onOutput?.(task.id, retryMessage);
-            try {
-              this.persistence.appendTaskOutput(task.id, retryMessage);
-            } catch {
-              // Preserve the original startup failure if output persistence also fails.
-            }
-            this.persistence.logEvent?.(task.id, 'task.executor.startup-retry', {
-              runnerKind: executor.type,
-              poolId: poolSelectionForStart.poolId,
-              poolMemberId: poolSelectionForStart.member.id,
-              reason: 'ssh-startup-transport-failure',
-              error: err instanceof Error ? err.message : String(err),
-            });
-            this.pendingPoolSelections.delete(task.id);
-            this.releasePoolSelectionLease(poolSelectionForStart);
-            continue;
-          }
-        }
-        const startupErrorMessage = `Executor startup failed (${executor.type}): ${err instanceof Error ? err.message : String(err)}\n`;
-        this.callbacks.onOutput?.(task.id, startupErrorMessage);
-        try {
-          this.persistence.appendTaskOutput(task.id, startupErrorMessage);
-        } catch {
-          // Preserve the original startup failure if output persistence also fails.
-        }
-        const launchStale = this.isLaunchStale(task.id, attemptId, startGeneration);
-        // Only persist startup-failure metadata when the launch is still
-        // current.  If the task has moved to a newer attempt or generation
-        // (e.g. via recreate-task), writing old workspace/branch metadata
-        // would corrupt the live attempt's state.
-        if (
-          (meta.workspacePath || meta.branch || meta.agentSessionId || meta.containerId)
-          && !launchStale
-        ) {
-          const execution: Record<string, string> = {};
-          if (meta.workspacePath) execution.workspacePath = meta.workspacePath;
-          if (meta.branch) execution.branch = meta.branch;
-          if (meta.agentSessionId) {
-            execution.agentSessionId = meta.agentSessionId;
-            execution.lastAgentSessionId = meta.agentSessionId;
-          }
-          if (meta.containerId) execution.containerId = meta.containerId;
-          const poolSelection = this.pendingPoolSelections.get(task.id);
-          const selectedSshTargetId = executor.type === 'ssh'
-            ? this.selectedRemoteTargetId(task, poolSelection)
-            : undefined;
-          this.persistence.updateTask(task.id, {
-            config: {
-              runnerKind: executor.type as RunnerKind,
-              ...(selectedSshTargetId ? { poolMemberId: selectedSshTargetId } : {}),
-            },
-            execution: execution as any,
-          });
-        }
-        if (launchStale) {
-          this.persistence.logEvent?.(task.id, 'task.executor.stale_startup_failure', {
-            attemptId,
-            executorType: executor.type,
-            error: err instanceof Error ? err.message : String(err),
-            workspacePath: meta.workspacePath,
-            branch: meta.branch,
-            hasAgentSessionId: Boolean(meta.agentSessionId),
-            hasContainerId: Boolean(meta.containerId),
-          });
-        }
-        this.pendingPoolSelections.delete(task.id);
-        this.releasePoolSelectionLease(poolSelectionForStart);
-        const wrapped = new Error(
-          `Executor startup failed (${executor.type}): ${err instanceof Error ? err.message : String(err)}`,
-          { cause: err },
-        );
-        if (!launchStale) {
-          this.callbacks.onLaunchFailed?.(task.id, wrapped, executor);
-        }
-        throw wrapped;
-      } finally {
-        clearInterval(preStartHeartbeatTimer);
-        if (preStartTimeout) clearTimeout(preStartTimeout);
-      }
-    }
-    traceExecution(`[trace] TaskRunner: task=${task.id} executor.start() returned after ${Date.now() - startT0}ms executor=${executor.type} sessionId=${handle.agentSessionId ?? 'none'} workspace=${handle.workspacePath ?? 'default'}`);
-    this.logger.info(
-      `[TaskRunner] executor.start returned task=${task.id} attempt=${attemptId} executor=${executor.type} ` +
-        `elapsedMs=${Date.now() - startT0} executionId=${handle.executionId} ` +
-        `workspace=${handle.workspacePath ?? 'none'} branch=${handle.branch ?? 'none'} ` +
-        `agentSessionId=${handle.agentSessionId ?? 'none'}`,
-    );
-    bench('executor.start.after', {
-      executorType: executor.type,
-      executorStartMs: Date.now() - startT0,
-      hasWorkspacePath: Boolean(handle.workspacePath),
-      hasAgentSessionId: Boolean(handle.agentSessionId),
+    const dispatched = await dispatchExecutor(this, {
+      task,
+      attemptId,
+      request,
+      bench,
+      dispatchOpts,
     });
-    const launchAccepted =
-      this.orchestrator.markTaskRunningAfterLaunch?.(task.id, attemptId) ?? true;
-    if (!launchAccepted) {
-      this.logger.warn(
-        `[TaskRunner] launch rejected as stale/non-executable for task=${task.id} attemptId=${attemptId}; killing spawned process`,
-      );
-      try {
-        await executor.kill(handle);
-      } catch (killErr) {
-        this.logger.warn(`[TaskRunner] failed to kill rejected launch for task=${task.id}`, { killErr });
-      }
-      this.releasePoolSelectionLease(this.pendingPoolSelections.get(task.id));
-      this.pendingPoolSelections.delete(task.id);
-      await this.cleanupPerTaskDockerExecutor(task);
-      if (dispatchOpts) {
-        dispatchOpts.launchOutbox.failDispatch(
-          dispatchOpts.dispatchId,
-          new Error('Launch rejected as stale or non-executable after executor start'),
-        );
-      }
-      bench('markTaskRunningAfterLaunch.rejected');
-      return;
-    }
-    bench('markTaskRunningAfterLaunch.accepted');
-
-    // Persist execution metadata immediately at task start — all fields explicit
-    {
-      // Fail-fast: workspacePath must be provided by all executors
-      if (!handle.workspacePath) {
-        this.releasePoolSelectionLease(this.pendingPoolSelections.get(task.id));
-        throw new Error(
-          `Executor "${executor.type}" did not provide workspacePath for task "${task.id}". ` +
-          `All executors must set workspacePath; refusing to fall back to host repo.`,
-        );
-      }
-
-      this.logExecutorSelected(
-        task,
-        executor,
-        handle,
-        attemptId,
-        this.pendingPoolSelections.get(task.id),
-      );
-
-      const poolSelection = this.pendingPoolSelections.get(task.id);
-      const selectedSshTargetId = executor.type === 'ssh'
-        ? this.selectedRemoteTargetId(task, poolSelection)
-        : undefined;
-      const changes = {
-        config: {
-          runnerKind: executor.type as RunnerKind,
-          ...(selectedSshTargetId ? { poolMemberId: selectedSshTargetId } : {}),
-        },
-        execution: {
-          workspacePath: handle.workspacePath,
-          branch: handle.branch ?? undefined,  // Explicit undefined when branch is not applicable (e.g., BYO mode)
-          agentSessionId: handle.agentSessionId ?? undefined,
-          lastAgentSessionId: handle.agentSessionId ?? undefined,
-          agentName: actionType === 'ai_task' ? executionAgent : undefined,
-          lastAgentName: actionType === 'ai_task' ? executionAgent : undefined,
-          containerId: handle.containerId ?? undefined,
-        },
-      };
-      this.persistence.updateTask(task.id, changes);
-      // Mirror branch + workspacePath onto the attempt row so reconciliation
-      // and post-mortem flows can recover provenance from the attempt without
-      // joining back to the task. Pairs with the early `onBranchResolved`
-      // persistence; this is the authoritative success-path write.
-      try {
-        this.persistence.updateAttempt?.(attemptId, {
-          branch: handle.branch ?? undefined,
-          workspacePath: handle.workspacePath,
-        } as any);
-      } catch (err) {
-        traceExecution(
-          `${RESTART_TO_BRANCH_TRACE} task=${task.id} attempt=${attemptId} post-start attempt persist failed: ${err instanceof Error ? err.message : String(err)}`,
-        );
-      }
-      traceExecution(
-        `[agent-session-trace] TaskRunner.persistStartMetadata task=${task.id} agentSessionId=${handle.agentSessionId ?? 'null'}`,
-      );
-      if (task.config.isMergeNode) {
-        traceExecution(
-          `[merge-gate-workspace] persistStartMetadata mergeNode=${task.id} ` +
-            `executor workspacePath=${changes.execution.workspacePath} ` +
-            '(gate clone path is written later in executeMergeNode)',
-        );
-      }
-      traceExecution(`[trace] TaskRunner: persisted metadata for task=${task.id} workspacePath=${handle.workspacePath} branch=${handle.branch ?? 'null'}`);
-      bench('persistStartMetadata.end', {
-        workspacePath: handle.workspacePath,
-        branch: handle.branch ?? undefined,
-      });
-    }
-
-    // Notify consumer about the spawned handle
-    const activeHandle = handle as ActiveExecutionHandle;
-    activeHandle.attemptId = attemptId;
-    const poolSelection = this.pendingPoolSelections.get(task.id);
-    this.pendingPoolSelections.delete(task.id);
-    this.activeExecutions.set(attemptId, {
-      handle: activeHandle,
-      executor,
-      taskId: task.id,
-      poolId: poolSelection?.poolId,
-      poolMemberKey: poolSelection?.memberKey,
-      leaseResourceKey: poolSelection?.leaseResourceKey,
-      leaseHolderId: poolSelection?.leaseHolderId,
-    });
-    this.logger.info(
-      `[TaskRunner] active execution registered task=${task.id} attempt=${attemptId} ` +
-        `executor=${executor.type} executionId=${handle.executionId} activeExecutions=${this.activeExecutions.size}`,
-    );
-    bench('onSpawned.before');
-    this.callbacks.onSpawned?.(task.id, handle, executor);
-    bench('onSpawned.after');
-
-    // Wire output
-    executor.onOutput(handle, (data) => {
-      this.callbacks.onOutput?.(task.id, data);
-    });
-
-    // Wire heartbeat
-    executor.onHeartbeat(handle, () => {
-      const now = new Date();
-      const isRemoteWorkloadHeartbeat = executor.type === 'ssh';
-      if (isRemoteWorkloadHeartbeat) {
-        this.logger.info(
-          `[TaskRunner] ssh heartbeat received task=${task.id} attempt=${attemptId} executionId=${handle.executionId} ` +
-            `at=${now.toISOString()}`,
-        );
-      }
-      const activeLease = this.activeExecutions.get(attemptId);
-      if (activeLease?.leaseResourceKey && activeLease.leaseHolderId) {
-        this.persistence.renewExecutionResourceLease?.(activeLease.leaseResourceKey, activeLease.leaseHolderId);
-      }
-      this.persistence.updateAttempt?.(attemptId, {
-        lastHeartbeatAt: now,
-        leaseExpiresAt: nextLeaseExpiry(now),
-      } as any);
-      this.callbacks.onHeartbeat?.(task.id, {
-        at: now,
-        source: isRemoteWorkloadHeartbeat ? 'remote_workload' : 'executor',
-      });
-    });
+    // Launch was rejected as stale/non-executable after executor start;
+    // dispatch already cleaned up and acked the dispatch row.
+    if (!dispatched) return;
+    const { executor, handle } = dispatched;
 
     // Wait for completion and feed response to orchestrator.
     // The callback is serialized through completionChain so that concurrent
